requirements.txt
psycopg2-binary==2.9.6
pandas==1.5.3